Move IO code to separate folder
