+++ src/Jellyfin.LiveTv/EmbyTV/EmbyTV.cs
@@ -19,6 +19,7 @@ using Jellyfin.Data.Enums;
 using Jellyfin.Data.Events;
 using Jellyfin.Extensions;
 using Jellyfin.LiveTv.Configuration;
+using Jellyfin.LiveTv.IO;
 using Jellyfin.LiveTv.Timers;
 using MediaBrowser.Common.Configuration;
 using MediaBrowser.Common.Extensions;
